requirements.txt
# Core dependencies
python-dotenv=0.19.0
py-cord=2.0.0
websockets=10.0
aiohttp=3.8.0
requests=2.25.0